requirements.txt
absl-py==0.1.10
appnope==0.1.0
backports.functools-lru-cache==1.5
bleach==2.0.0
certifi==2018.1.18
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
h5py==2.7.1
html5lib==1.0.1
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.4
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.1
olefile==0.45.1
pandas==0.22.0
pandocfilters==1.4.1
parso==0.1.1
pbr==3.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
pip==9.0.1
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
scipy==1.0.0
Send2Trash==1.5.0
setuptools==38.5.1
simplegeneric==0.8.1
six==1.11.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.14.1
wheel==0.30.0
widgetsnbextension==3.1.4
